reuse ruleset.Run across sets
